+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Bug Fixes
 
-* remove stale @duraflows/core@0.0.2 nested in nestjs package ([60d3682](https://github.com/camcima/duraflows/commit/60d36828efc953f6a3a6efcd400dd9cc08e8319a))
+- remove stale @duraflows/core@0.0.2 nested in nestjs package ([60d3682](https://github.com/camcima/duraflows/commit/60d36828efc953f6a3a6efcd400dd9cc08e8319a))
 
 ## [0.1.0](https://github.com/camcima/duraflows/compare/v0.0.2...v0.1.0) (2026-03-30)